+++ packages/styled-components/components/socialLinks/index.js
@@ -45,8 +45,8 @@ const SocialLinks = (props) => {
           {items.map(({ platform, url, icon }) => (
             <SocialLinksItem key={platform}>
               <Link href={url}>
-                <IconWrapper>
-                  {icon.type.displayName ? icon : platform}
+                <IconWrapper className={platform}>
+                  {icon}
                 </IconWrapper>
               </Link>
             </SocialLinksItem>
@@ -66,7 +66,7 @@ SocialLinks.defaultProps = {
 SocialLinks.propTypes = {
   ...standardPropTypes,
   /**
-   * An object containing social platforms as keys as links to specific pages as values.
+   * An object containing social platforms as keys and links to specific pages as values.
    */
   platforms: PropTypes.oneOfType([
     PropTypes.object,