+++ CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.3.0-beta.3](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.2...v3.3.0-beta.3) (2020-08-31)
+
+
+### Bug Fixes
+
+* **sociallinks:** fix icon display issue, add class to icon wrapper, and tweak CSS for social links ([132f463](https://github.com/alleyinteractive/irving/commit/132f46320224618b087305f95fe9782a9f6f9382))
+
+
+### Features
+
+* **socialsharing:** allow for custom links in Social Sharing component and refactor to one file ([54f36b3](https://github.com/alleyinteractive/irving/commit/54f36b36f330876540789bbaf32725136765403d))
+
+
+
+
+
+
+
+**Note:** Version bump only for package irving
+
+
+
+
+
+# [3.3.0-beta.2](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.1...v3.3.0-beta.2) (2020-08-27)
+
+
+### Bug Fixes
+
+* **commentcount:** fix reference ([3c5e554](https://github.com/alleyinteractive/irving/commit/3c5e5546730cfb0e2a5171931bbbb46ae5ed7c6f))
+
+
+
+
+
+# [3.3.0-beta.1](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.0...v3.3.0-beta.1) (2020-08-27)
+
+
+### Bug Fixes
+
+* **cacheclient:** return default client, not service, from VIP go cacheClient ([07a2249](https://github.com/alleyinteractive/irving/commit/07a2249db017f9c998b11e0cccd73497681738d4))
+* **Coral:** Fix client rendering activation of comment counts ([3a48963](https://github.com/alleyinteractive/irving/commit/3a489633f1fea785ccfe32f75b1e11057636a97b))
+
+
+
+
+
+# [3.3.0-beta.0](https://github.com/alleyinteractive/irving/compare/v3.3.0-alpha.1...v3.3.0-beta.0) (2020-08-27)
+
+**Note:** Version bump only for package irving
+
+
+
+
+
 ## [3.2.1](https://github.com/alleyinteractive/irving/compare/v3.2.1-alpha.0...v3.2.1) (2020-08-21)